+++ CHANGELOG.md
@@ -18,6 +18,8 @@ New features:
 - Add several volume control channels (TODO websocket commands for these).
 - Change Loudness filter to only perform loudness compensation.
 - Add Triangular and Gaussian dithers.
+- Add Short and Long variants of Fweighted441 dithers.
+- Add Long variant of Lipshitz441 dither.
 
 Changes:
 - Optimize dithering cpu load.